cli: move more self update logic into self_update module
